Let oryx app running in three containers.
three containers running in a single pod. The three containers are oryx
platform(including oryx backend and front-end WebUI), redis and srs.
redis and srs are sidecar container, the oryx is the application
container.
The purpose of isolated containers is to support livenessProbe and
readinessProbe for each processes, oryx platform, redis, srs.